+++ packages/components/src/toggle-control/test/index.tsx
@@ -1,11 +1,4 @@
-/**
- * External dependencies
- */
 import { render, screen } from '@testing-library/react';
-
-/**
- * Internal dependencies
- */
 import ToggleControl from '..';
 
 describe( 'ToggleControl', () => {