+++ packages/editor/src/components/local-autosave-monitor/index.js
@@ -1,16 +1,9 @@
-/**
- * WordPress dependencies
- */
 import { useCallback, useEffect, useMemo, useRef } from '@wordpress/element';
 import { ifCondition, usePrevious } from '@wordpress/compose';
 import { useSelect, useDispatch, useRegistry } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 import { parse } from '@wordpress/blocks';
 import { store as noticesStore } from '@wordpress/notices';
-
-/**
- * Internal dependencies
- */
 import AutosaveMonitor from '../autosave-monitor';
 import {
 	localAutosaveGet,