+++ packages/editor/src/components/post-schedule/test/check.js
@@ -1,16 +1,5 @@
-/**
- * External dependencies
- */
 import { render, screen } from '@testing-library/react';
-
-/**
- * WordPress dependencies
- */
 import { useSelect } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
 import PostScheduleCheck from '../check';
 
 jest.mock( '@wordpress/data/src/components/use-select', () => jest.fn() );