+++ packages/wp-build/CHANGELOG.md
@@ -57,9 +57,11 @@
 
 -   Fix `genereated` typo to `generated` in README. ([#79331](https://github.com/WordPress/gutenberg/pull/79331))
 
-## 0.16.1 (2026-06-16)
+### Enhancements
 
-## 0.16.0 (2026-06-10)
+-   Restore optimization to exclude experimental block-library code from WordPress core bundles via a built-in ESBuild plugin ([#79306](https://github.com/WordPress/gutenberg/pull/79306)).
+
+## 0.16.1 (2026-06-16)
 
 ### Bug Fixes